+++ packages/core/badge/docs/0-intro.js
@@ -1,7 +1,7 @@
 // @flow
 
 import React from 'react';
-import { md, Example, Props } from '@atlaskit/docs';
+import { code, md, Example, Props } from '@atlaskit/docs';
 
 export default md`
   Badges are visual indicators for numeric values such as tallies and scores.
@@ -14,19 +14,11 @@ export default md`
   * Use labels to call out tags and high-visibility attributes.
   * Use a tooltip if you want to indicate units.
 
-  ## Installation
-
-  \`\`\`sh
-  yarn add @atlaskit/badge
-  \`\`\`
-
-  Interact with a [live demo of the @NAME@ component with code examples](https://aui-cdn.atlassian.com/atlaskit/stories/@NAME@/@VERSION@/).
-
   ## Usage
 
   The \`default\` export gives you full badge functionality and automatically formats the number you priovide it.
 
-  \`\`\`js
+${code`
   import Badge from '@atlaskit/badge';
 
   // Displays: 99+
@@ -34,33 +26,32 @@ export default md`
 
   // Displays: 999+
   <Badge max={999}>{1000}</Badge>
-  \`\`\`
+`}
 
-  ### Container
+  ## Container
 
   The named \`Container\` export retains the styling of a normal badge, but without formatting. This means you can compose in whatever information you need to.
 
-  \`\`\`js
+${code`
   import { Container } from '@atlaskit/badge';
 
   // Displays: <em>Something</em>
   <Container><em>Something</em></Container>
-  \`\`\`
+
+`}
 
   _Beware that putting arbitrary content inside of a badge might cause it to take on an unitended look._
 
-  ### Format
+  ## Format
 
-  The \`Format\` export can be used to compose your own badge together, or if you need the badge  style formatting somewhere else.
+  The \`Format\` export can be used to compose your own badge together, or if you need the badge style formatting somewhere else.
 
-  \`\`\`js
+  ${code`
   import { Container, Format } from '@atlaskit/badge';
 
   // Displays: <em>999+</em>
   <Container><em><Format>{1000}</Format></em></Container>
-  \`\`\`
-
-  ## Examples
+`}
 
   ${(
     <Example
@@ -70,5 +61,10 @@ export default md`
     />
   )}
 
-  ${<Props props={require('!!extract-react-types-loader!../src')} />}
+  ${(
+    <Props
+      heading="Badge Props"
+      props={require('!!extract-react-types-loader!../src/components')}
+    />
+  )}
 `;