+++ packages/core/datetime-picker/src/components/DatePicker.js
@@ -3,7 +3,7 @@
 import Calendar from '@atlaskit/calendar';
 import CalendarIcon from '@atlaskit/icon/glyph/calendar';
 import Select, { mergeStyles } from '@atlaskit/select';
-import { borderRadius, colors, layers } from '@atlaskit/theme';
+import { borderRadius, colors, layers, elevation } from '@atlaskit/theme';
 import { format, isValid, parse } from 'date-fns';
 import pick from 'lodash.pick';
 import React, { Component, type Node, type ElementRef } from 'react';
@@ -52,6 +52,8 @@ type Props = {
   /** Props to apply to the select. This can be used to set options such as placeholder text.
    *  See [here](/packages/core/select) for documentation on select props. */
   selectProps: Object,
+  /* This prop affects the height of the select control. Compact is gridSize() * 4, default is gridSize * 5  */
+  spacing?: 'compact' | 'default',
   /** The ISO time that should be used as the input value. */
   value?: string,
   /** Indicates current value is invalid & changes border color */
@@ -72,21 +74,6 @@ type State = {
   view: string,
 };
 
-// TODO see if there's a different way to control the display value.
-//
-// react-select retains the value the user typed in until the field is
-// blurred. Since we're controlling the open state and value, we need a
-// way explicitly ensure the value is respected. By blurring and then
-// immedately refocusing, we ensure the value is formatted and the input
-// retains focus.
-function ensureValueIsDisplayed() {
-  const { activeElement } = document;
-  if (activeElement) {
-    activeElement.blur();
-    activeElement.focus();
-  }
-}
-
 function isoToObj(iso: string) {
   const parsed = parse(iso);
   return isValid(parsed)
@@ -106,11 +93,10 @@ const arrowKeys = {
 };
 
 const StyledMenu = styled.div`
-  background-color: ${colors.N0};
-  border: 1px solid ${colors.N40};
+  background-color: ${colors.N20};
+  border: 0 0 1px solid ${colors.N60A};
   border-radius: ${borderRadius()}px;
-  box-shadow: 1px 5px 10px rgba(0, 0, 0, 0.1);
-  margin: 7px 0;
+  ${elevation.e200} margin: 8px 0 0 0;
   overflow: hidden;
   text-align: center;
   z-index: ${layers.dialog};
@@ -134,6 +120,7 @@ export default class DatePicker extends Component<Props, State> {
     onFocus: () => {},
     innerProps: {},
     selectProps: {},
+    spacing: 'default',
     id: '',
     defaultIsOpen: false,
     defaultValue: '',
@@ -164,12 +151,12 @@ export default class DatePicker extends Component<Props, State> {
   };
 
   onCalendarSelect = ({ iso: value }: { iso: string }) => {
-    this.triggerChange(value);
     this.setState({ isOpen: false, selectedValue: value });
+    this.triggerChange(value);
   };
 
   onInputClick = () => {
-    this.setState({ isOpen: true });
+    if (!this.getState().isOpen) this.setState({ isOpen: true });
   };
 
   onSelectBlur = (e: SyntheticFocusEvent<>) => {
@@ -230,7 +217,6 @@ export default class DatePicker extends Component<Props, State> {
   triggerChange = (value: string) => {
     this.props.onChange(value);
     this.setState({ value, view: value });
-    ensureValueIsDisplayed();
   };
 
   getContainerRef = (ref: ?HTMLElement) => {
@@ -262,6 +248,7 @@ export default class DatePicker extends Component<Props, State> {
       isDisabled,
       name,
       selectProps,
+      spacing,
       dateFormat,
       placeholder,
     } = this.props;
@@ -279,7 +266,7 @@ export default class DatePicker extends Component<Props, State> {
           disabled={disabled}
           onChange={this.onCalendarChange}
           onSelect={this.onCalendarSelect}
-          // $FlowFixMe
+          // $FlowFixMe - Calendar is not a react component
           ref={this.refCalendar}
           selected={[this.state.selectedValue]}
           innerProps={menuInnerProps}
@@ -296,7 +283,7 @@ export default class DatePicker extends Component<Props, State> {
     const { styles: selectStyles = {} } = selectProps;
     const controlStyles =
       this.props.appearance === 'subtle' ? this.getSubtleControlStyles() : {};
-
+    const disabledStyle = isDisabled ? { pointerEvents: 'none' } : {};
     return (
       <div
         {...innerProps}
@@ -309,7 +296,7 @@ export default class DatePicker extends Component<Props, State> {
         <input name={name} type="hidden" value={value} />
         {/* $FlowFixMe - complaining about required args that aren't required. */}
         <Select
-          menuIsOpen={isOpen}
+          menuIsOpen={isOpen && !isDisabled}
           openMenuOnFocus
           closeMenuOnSelect
           autoFocus={autoFocus}
@@ -326,6 +313,7 @@ export default class DatePicker extends Component<Props, State> {
             control: base => ({
               ...base,
               ...controlStyles,
+              ...disabledStyle,
             }),
           })}
           placeholder={placeholder}
@@ -336,6 +324,7 @@ export default class DatePicker extends Component<Props, State> {
             }
           }
           {...selectProps}
+          spacing={spacing}
           validationState={validationState}
         />
       </div>