+++ packages/core/field-text/src/FieldText.js
@@ -38,7 +38,6 @@ export default class FieldText extends Component<FieldTextProps, State> {
         value={this.state.value}
         onChange={this.handleOnChange}
         ref={fieldRef => {
-          // $FlowFixMe - `FieldTextStateless` is incompatible with `HTMLInputElement`
           this.input = fieldRef;
         }}
       />