+++ packages/core/icon/utils/icons.js
@@ -77,6 +77,7 @@ import CrossCircleIcon from '../glyph/cross-circle';
 import CrossIcon from '../glyph/cross';
 import DashboardIcon from '../glyph/dashboard';
 import DecisionIcon from '../glyph/decision';
+import DepartmentIcon from '../glyph/department';
 import DetailViewIcon from '../glyph/detail-view';
 import DiscoverFilledIcon from '../glyph/discover-filled';
 import DiscoverIcon from '../glyph/discover';
@@ -372,6 +373,7 @@ import ScheduleFilledIcon from '../glyph/schedule-filled';
 import ScheduleIcon from '../glyph/schedule';
 import ScreenIcon from '../glyph/screen';
 import SearchIcon from '../glyph/search';
+import SelectClearIcon from '../glyph/select-clear';
 import SendIcon from '../glyph/send';
 import SettingsIcon from '../glyph/settings';
 import ShareIcon from '../glyph/share';
@@ -384,6 +386,7 @@ import StarIcon from '../glyph/star';
 import StatuspageIcon from '../glyph/statuspage';
 import StrideIcon from '../glyph/stride';
 import SubtaskIcon from '../glyph/subtask';
+import SuitcaseIcon from '../glyph/suitcase';
 import SwitcherIcon from '../glyph/switcher';
 import TableIcon from '../glyph/table';
 import TaskIcon from '../glyph/task';
@@ -827,6 +830,12 @@ export default {
     component: DecisionIcon,
     componentName: 'DecisionIcon',
     package: '@atlaskit/icon/glyph/decision',
+  },
+  department: {
+    keywords: ['department'],
+    component: DepartmentIcon,
+    componentName: 'DepartmentIcon',
+    package: '@atlaskit/icon/glyph/department',
   },
   'detail-view': {
     keywords: ['detail-view', 'detailview'],
@@ -2598,6 +2607,12 @@ export default {
     componentName: 'SearchIcon',
     package: '@atlaskit/icon/glyph/search',
   },
+  'select-clear': {
+    keywords: ['select-clear', 'selectclear'],
+    component: SelectClearIcon,
+    componentName: 'SelectClearIcon',
+    package: '@atlaskit/icon/glyph/select-clear',
+  },
   send: {
     keywords: ['send', 'mail'],
     component: SendIcon,
@@ -2669,6 +2684,12 @@ export default {
     component: SubtaskIcon,
     componentName: 'SubtaskIcon',
     package: '@atlaskit/icon/glyph/subtask',
+  },
+  suitcase: {
+    keywords: ['suitcase'],
+    component: SuitcaseIcon,
+    componentName: 'SuitcaseIcon',
+    package: '@atlaskit/icon/glyph/suitcase',
   },
   switcher: {
     keywords: ['switcher'],