+++ packages/core/tree/mockdata/TreeBuilder.js
@@ -1,7 +1,6 @@
 //@flow
 
 import type { ItemId, TreeItem } from '../src/types';
-import { mutateTreeItem } from '../src/utils/tree';
 
 export default class TreeBuilder {
   rootId: ItemId;
@@ -27,12 +26,13 @@ export default class TreeBuilder {
     this._addItemToRoot(`${this.rootId}-${subTree.rootId}`);
     Object.keys(subTree.items).forEach(itemId => {
       const finalId = `${this.rootId}-${itemId}`;
-      this.items[finalId] = mutateTreeItem(subTree.items[itemId], {
+      this.items[finalId] = {
+        ...subTree.items[itemId],
         id: finalId,
         children: subTree.items[itemId].children.map(
           i => `${this.rootId}-${i}`,
         ),
-      });
+      };
     });
     return this;
   }