+++ packages/core/tree/src/components/Tree/Tree-types.js
@@ -1,9 +1,9 @@
 // @flow
 import * as React from 'react';
-import type { TreeData, Path, ItemId } from '../../types';
+import type { TreeData, Path, ItemId, FlattenedTree } from '../../types';
 import { type RenderItemParams } from '../TreeItem/TreeItem-types';
 
-export type DragPosition = {|
+export type TreePosition = {|
   parentId: ItemId,
   index: number,
 |};
@@ -14,14 +14,15 @@ export type Props = {|
   onCollapse: (itemId: ItemId, path: Path) => void,
   onDragStart: (itemId: ItemId) => void,
   onDragEnd: (
-    sourcePosition: DragPosition,
-    destinationPosition: DragPosition,
+    sourcePosition: TreePosition,
+    destinationPosition: ?TreePosition,
   ) => void,
   renderItem: RenderItemParams => React.Node,
-  paddingPerLevel: number,
+  offsetPerLevel: number,
   isDragEnabled: boolean,
 |};
 
 export type State = {|
+  flattenedTree: FlattenedTree,
   dropAnimationOffset: number,
 |};